+++ CHANGES.md
@@ -88,11 +88,14 @@ Version PBFT
  -  (Libplanet.Net) Added `Proposal` class.  [[#PBFT]]
  -  (Libplaent.Net) Added `ProposalMetadata` class.  [[#PBFT]]
  -  (Libplanet.Explorer) Added `BoundPeerType` class.  [[#PBFT]]
+ -  (Libplanet.Explorer) Added `BlockCommitType` class.  [[#PBFT]]
  -  (Libplanet.Explorer) Added `VoteFlagType` class.  [[#PBFT]]
  -  (Libplanet.Explorer) Added `VoteType` class.  [[#PBFT]]
  -  (Libplanet.Explorer) Added `VoteCount` class.  [[#PBFT]]
  -  (Libplanet.Explorer) Added `VoteCountType` class.  [[#PBFT]]
  -  (Libplanet.Explorer) Added `VoteQuery` class.  [[#PBFT]]
+ -  (Libplanet.Explorer) Added `BlockCommitType` as a return of `BlockQuery`.
+    [[#PBFT]]
 
 ### Behavioral changes